backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "fastapi>=0.115.0",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "uvicorn>=0.32.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0"
]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.2",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "black>=24.10.0",
    "flake8>=7.1.1",
    "mypy>=1.13.0"
]
